requirements.txt
flask
google-generativeai
python-telegram-bot
gunicorn
pillow